Beta

+++ readarr/rootfs/etc/cont-init.d/20-folders.sh
@@ -1,9 +1,9 @@
-#!/usr/bin/with-contenv bashio
+#!/bin/bash
 
-if [ ! -d /share/music ]; then
-  echo "Creating /share/music"
-  mkdir -p /share/music
-  chown -R abc:abc /share/music
+if [ ! -d /share/storage/ebook ]; then
+  echo "Creating /share/storage/ebook"
+  mkdir -p /share/storage/ebook
+  chown -R abc:abc /share/storage/ebook
 fi
 
 if [ ! -d /share/downloads ]; then
@@ -12,15 +12,16 @@ if [ ! -d /share/downloads ]; then
   chown -R abc:abc /share/downloads
 fi
 
-if [ ! -d /config/addons_config/readarr ]; then
-if [ -d /config/readarr ]; then
+if [ -d /config/readarr ] && [ ! -d /config/addons_config/readarr ]; then
   echo "Moving to new location /config/addons_config/readarr"
   mkdir -p /config/addons_config/readarr
   chown -R abc:abc /config/addons_config/readarr
   mv /config/readarr/* /config/addons_config/readarr/
   rm -r /config/readarr
-fi 
- echo "Creating /config/addons_config/readarr"
+fi
+
+if [ ! -d /config/addons_config/readarr ]; then
+  echo "Creating /config/addons_config/readarr"
   mkdir -p /config/addons_config/readarr
   chown -R abc:abc /config/addons_config/readarr
 fi
